multicaller: Implement balance_call
